Fix bugs and add tests

+++ samgraph/common/cuda_sampling.h
@@ -10,7 +10,7 @@ namespace common {
 namespace cuda {
 
 void DeviceSample(const nodeid_t *indptr, const nodeid_t *indices, const size_t num_node, const size_t num_edge, const nodeid_t *input,
-                  const size_t num_input, const int fanout, nodeid_t *out_src, nodeid_t *out_dst, size_t *num_out, cudaStream_t stream);
+                  const size_t num_input, const size_t fanout, nodeid_t *out_src, nodeid_t *out_dst, size_t *num_out, cudaStream_t stream);
 
 } // namespace cuda
 } // namespace common
